Transactions, customer migrations
I'm not 100% convinced I defined the datatypes correctly for our
customers, transactions, and merchants models.

That said, this was partially intentional, as I wanted practice making
regular and minor changes to our schemas.

So, since we don't have any relationships between the models, there's a
column for some foreign keys, but no actual data.

I look forward to resolving this.

Our rake tasks feel clunky and wrong, but it'll do for now.